New physics manager works with old methods

+++ UnityDLL/UnityDLL/Fixer.h
@@ -5,17 +5,17 @@ class Fixer
 {
 
 public:
-	Fixer() : position(Eigen::Vector3f(0, 0, 0)), scale(Eigen::Vector3f(1, 1, 1)) {}
+	Fixer() : position(Eigen::Vector3d(0, 0, 0)), scale(Eigen::Vector3d(1, 1, 1)) {}
 	Fixer(Vector3f _position, Vector3f _scale) {
-		position = Eigen::Vector3f(_position.x, _position.y, _position.z);
-		scale = Eigen::Vector3f(_scale.x, _scale.y, _scale.z);
+		position = Eigen::Vector3d(_position.x, _position.y, _position.z);
+		scale = Eigen::Vector3d(_scale.x, _scale.y, _scale.z);
 	}
 	~Fixer() { nodesInside.clear(); }
 
 	bool CheckNodeInside(Node* n);
 
-	Eigen::Vector3f position;
-	Eigen::Vector3f scale = Eigen::Vector3f(1, 1, 1);
+	Eigen::Vector3d position;
+	Eigen::Vector3d scale = Eigen::Vector3d(1, 1, 1);
 
 	std::vector<Node*> nodesInside;
 };
